Bug Fix for network emits
Originally the address used for the SSID would become inconsistent because of the way the ESP32 uses the string class to store the SSID.  This would eventually end in an emit that crashed the socket from the client with an inconsistent UTF-8 byte.  The fix was to convert the string to a pointer for the emits.